Pull out python_unittests_unix.py
The path tests can only work on UNIX-like OSs because Python's path
joining is required to use / as the path separator
