+++ src/scripts/python_uniitests.py
@@ -10,14 +10,13 @@ Requires Python 3.
 Botan is released under the Simplified BSD License (see license.txt)
 """
 
-import os
 import sys
 import unittest
 
 sys.path.append("../..") # Botan repo root
 from configure import CompilerDetector # pylint: disable=wrong-import-position
 from configure import ModulesChooser # pylint: disable=wrong-import-position
-from install import prepend_destdir # pylint: disable=wrong-import-position
+
 
 class CompilerDetection(unittest.TestCase):
 
@@ -268,34 +267,6 @@ class ModulesChooserResolveDependencies(unittest.TestCase):
         ok, modules = ModulesChooser.resolve_dependencies(available_modules, table, "G")
         self.assertTrue(ok)
         self.assertEqual(modules, set(["G", "A", "C", "E"]))
-
-
-class PrependDestdir(unittest.TestCase):
-    def test_base(self):
-        os.environ["DESTDIR"] = "/"
-        self.assertEqual(prepend_destdir("/home/me"), "/home/me")
-        self.assertEqual(prepend_destdir("relative_path"), "/relative_path")
-        self.assertEqual(prepend_destdir("./relative_path"), "/relative_path")
-        self.assertEqual(prepend_destdir("relative/sub"), "/relative/sub")
-
-        self.assertEqual(prepend_destdir("/home/me/"), "/home/me")
-        self.assertEqual(prepend_destdir("relative_path/"), "/relative_path")
-
-        self.assertEqual(prepend_destdir("/home/me/../me2"), "/home/me2")
-        self.assertEqual(prepend_destdir("relative/sub/../sub2"), "/relative/sub2")
-
-        os.environ["DESTDIR"] = "/opt"
-        self.assertEqual(prepend_destdir("/home/me"), "/opt/home/me")
-        self.assertEqual(prepend_destdir("relative_path"), "/opt/relative_path")
-        self.assertEqual(prepend_destdir("./relative_path"), "/opt/relative_path")
-        self.assertEqual(prepend_destdir("relative/sub"), "/opt/relative/sub")
-
-        self.assertEqual(prepend_destdir("/home/me/"), "/opt/home/me")
-        self.assertEqual(prepend_destdir("relative_path/"), "/opt/relative_path")
-
-        self.assertEqual(prepend_destdir("/home/me/../me2"), "/opt/home/me2")
-        self.assertEqual(prepend_destdir("relative/sub/../sub2"), "/opt/relative/sub2")
-
 
 
 if __name__ == '__main__':
